move the logger to the base class

+++ simulator/heartbeat.py
@@ -1,11 +1,11 @@
 import time
-import paho.mqtt.client as mqtt
 import argparse
+import paho.mqtt.client as mqtt
 
 MQTT_BROKER = "localhost"
 
 parser = argparse.ArgumentParser(description="Heartbeat Publisher")
-parser.add_argument("interval", type=int, help="Heartbeat interval in seconds")
+parser.add_argument("interval", type=float, help="Heartbeat interval in seconds")
 args = parser.parse_args()
